Refactored file
