+++ src/backend/api/Fusion.Resources.Api/Controllers/Requests/ApiModels/ApiResourceAllocationRequest.cs
@@ -94,22 +94,4 @@ namespace Fusion.Resources.Api.Controllers
         public IEnumerable<ApiRequestComment>? Comments { get; set; }
         public ApiProvisioningStatus ProvisioningStatus { get; set; }
     }
-
-    public class ApiProposalParameters
-    {
-        public ApiProposalParameters(QueryResourceAllocationRequest.QueryPropsalParameters proposalParameters)
-        {
-            ChangeDateFrom = proposalParameters.ChangeFrom;
-            ChangeDateTo = proposalParameters.ChangeTo;
-
-            Scope = proposalParameters.Scope;
-            Type = proposalParameters.ChangeType;
-        }
-
-        public DateTime? ChangeDateFrom { get; set; }
-        public DateTime? ChangeDateTo { get; set; }
-
-        public string Scope { get; set; }
-        public string? Type { get; set; }
-    }
 }
